tests: (Hack) Added hack to timeseries_data test in data_management
Motor and pytest_asyncio don't play nice here running in different
event_loops (motor seems to generate their own depending on the context
motor functions are called).
To fix this the signal_bucket has to be created in the async function
instead of the fixture in conftest.
This solution is considered a hack.